+++ backend/canisters/group/api/src/lib.rs
@@ -28,6 +28,16 @@ pub enum LocalIndexEvent {
     UserDeleted(UserId),
     BotUpdated(BotDefinitionUpdate),
     BotRemoved(UserId),
+    UserIdMigrated(UserIdMigrated),
+}
+
+// A user has been migrated to a MultiUser canister, which gave them a new id
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct UserIdMigrated {
+    #[serde(rename = "o")]
+    pub old_user_id: UserId,
+    #[serde(rename = "n")]
+    pub new_user_id: UserId,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]